move store setup to standalone function
